requirements.txt
flask
Flask-SQLAlchemy
flask-migrate
flask-restful
qrcode
Pillow